+++ src/Data/Tuple/Constraint.hs
@@ -10,6 +10,9 @@
 -- | This module provides classes that emulate the behavior of GHC's constraint
 -- tuple syntax. Unlike GHC's built-in constraint tuples, the classes in this
 -- library can be partially applied.
+--
+-- When building with GHC 9.10 or later, this will simply re-export the
+-- constraint tuples offered by "GHC.Classes".
 module Data.Tuple.Constraint
   ( -- * Constraint tuples
     CTuple0
@@ -85,7 +88,13 @@ module Data.Tuple.Constraint
 #endif
   ) where
 
+#if __GLASGOW_HASKELL__ >= 910
+import GHC.Classes
+#else
 import Data.Kind (Constraint)
+#endif
+
+#if __GLASGOW_HASKELL__ < 910
 
 -- | An alias for a nullary constraint tuple.
 type CTuple0 = (() :: Constraint)
@@ -357,3 +366,4 @@ class    (c1, c2, c3, c4, c5, c6, c7, c8, c9, c10, c11, c12, c13, c14, c15, c16,
 instance (c1, c2, c3, c4, c5, c6, c7, c8, c9, c10, c11, c12, c13, c14, c15, c16, c17, c18, c19, c20, c21, c22, c23, c24, c25, c26, c27, c28, c29, c30, c31, c32, c33, c34, c35, c36, c37, c38, c39, c40, c41, c42, c43, c44, c45, c46, c47, c48, c49, c50, c51, c52, c53, c54, c55, c56, c57, c58, c59, c60, c61, c62, c63, c64) => CTuple64 c1 c2 c3 c4 c5 c6 c7 c8 c9 c10 c11 c12 c13 c14 c15 c16 c17 c18 c19 c20 c21 c22 c23 c24 c25 c26 c27 c28 c29 c30 c31 c32 c33 c34 c35 c36 c37 c38 c39 c40 c41 c42 c43 c44 c45 c46 c47 c48 c49 c50 c51 c52 c53 c54 c55 c56 c57 c58 c59 c60 c61 c62 c63 c64
 #endif
 
+#endif