+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## next [????.??.??]
 * Drop support for GHC 7.10 and earlier.
+* When building with GHC 9.10 or later, `Data.Tuple.Constraint` will simply
+  re-export constraint tuple definitions found in `GHC.Classes`. If building
+  with an older version of GHC, `Data.Tuple.Constraint` will backport the
+  definitions from `GHC.Classes.
 * Provide nullary constraint tuples named `CUnit` and unary constraint tuples
   named `CSolo` for consistency with `GHC.Classes` in GHC 9.10+.
 * Provide constraint tuples up to size 64 when building with GHC 9.2 or later.